Incorporating PR feedback
