+++ client/manipulators/placement.hpp
@@ -1,16 +1,13 @@
 #pragma once
 
-#include <algorithm>
-
-#include "engine-entities.hpp"
-#include "graphics/renderable.hpp"
-#include "resources/mesh.hpp"
+#include <memory>
+#include <glm/glm.hpp>
 
 namespace tec {
-using RenderableMap = Multiton<eid, Renderable*>;
+class MeshFile;
 namespace manipulator {
 
-const float MAX_DISTANCE = 25.0f; // Maximum distance at which to place.
+const float DEFAULT_MAX_PLACEMENT_DISTANCE = 25.0f;
 
 /**
 * \brief Manipulator used to place objects in the world. Can have a mesh assigned to provide visual feedback.
@@ -18,35 +15,19 @@ const float MAX_DISTANCE = 25.0f; // Maximum distance at which to place.
 */
 class Placement {
 public:
+	void SetMaxDistance(float _max_distance = DEFAULT_MAX_PLACEMENT_DISTANCE);
 	/**
 	* \brief Set the mesh used to represent the placement in the world.
 	* 
 	* \param[in] const std::shared_ptr<MeshFile> mesh MeshFile to render in the world.
 	*/
-	void SetMesh(const std::shared_ptr<MeshFile> _mesh) {
-		auto renderable = GetRenderable();
-		if (renderable) {
-			this->ClearMesh();
-			renderable->mesh = _mesh;
-			this->mesh = _mesh;
-		}
-	}
+	void SetMesh(const std::shared_ptr<MeshFile> _mesh);
 
 	/**
 	* \brief Clear the mesh, effectively not longer rendering the placement.
 	* 
 	*/
-	void ClearMesh() {
-		auto renderable = GetRenderable();
-		if (renderable) {
-			if (renderable->buffer) {
-				renderable->buffer->Destroy();
-				renderable->buffer.reset();
-			}
-			renderable->mesh.reset();
-			this->mesh.reset();
-		}
-	}
+	void ClearMesh();
 
 	/**
 	* \brief Sets the ray used to render the placement.
@@ -57,36 +38,14 @@ public:
 	* \param[in] const glm::vec3 start Starting point for the ray, usually the center of the screen.
 	* \param[in] const glm::vec3 intersection Intersection point of the ray with the physics world.
 	*/
-	void SetRayIntersectionPoint(const glm::vec3 start, const glm::vec3 intersection) {
-		auto renderable = GetRenderable();
-		float distance = glm::distance(start, intersection);
-		glm::vec3 direction = glm::normalize(intersection - start);
-		renderable->local_translation.value = start + direction * std::min<float>(distance, MAX_DISTANCE);
-	}
+	void SetRayIntersectionPoint(const glm::vec3 start, const glm::vec3 intersection);
 
-	void PlaceEntityInWorld(glm::vec3 _position) {
-		if (this->mesh) {
-
-			static eid starting_entity_id = 20000;
-			proto::Entity entity;
-			entity.set_id(starting_entity_id++);
-
-			Renderable renderable;
-			renderable.mesh_name = this->mesh->GetName();
-			renderable.Out(entity.add_components());
-
-			Position position(_position);
-			position.Out(entity.add_components()->mutable_position());
-
-			EventSystem<EntityCreated>::Get()->Emit(std::make_shared<EntityCreated>(EntityCreated{entity}));
-
-			this->ClearMesh();
-		}
-	}
+	void PlaceEntityInWorld(glm::vec3 _position);
 
 private:
-	Renderable* GetRenderable() { return RenderableMap::Get(ENGINE_ENTITIES::MANIPULATOR); }
 	std::shared_ptr<MeshFile> mesh;
+
+	float max_distance = DEFAULT_MAX_PLACEMENT_DISTANCE; // Maximum distance at which to place.
 };
 } // namespace manipulator
 } // namespace tec
